Take a list of integers when setting the fiducial ID filter Fiducial IDs are always integers, so this just takes a list of integers instead. I think this was actually the original intention too, because of the way that it streams the list through a mapping function, which isn't needed to just get a double array from a double list, but I guess it just never got implemented properly for some reason (I actually suspect that the Limelight might take an integer array already and this whole step is unnecessary, since their docs have a tendency to be a bit unreliable, but I need to do some more testing to be sure).

+++ src/main/java/io/github/roboblazers7617/limelight/LimelightSettings.java
@@ -299,8 +299,10 @@ public class LimelightSettings {
 	 * @return
 	 *         This object for method chaining.
 	 */
-	public LimelightSettings withAprilTagIdFilter(List<Double> idFilter) {
-		fiducialIDFiltersOverride.set(idFilter.stream().mapToDouble(Double::doubleValue).toArray());
+	public LimelightSettings withAprilTagIdFilter(List<Integer> idFilter) {
+		fiducialIDFiltersOverride.set(idFilter.stream()
+				.mapToDouble(Integer::doubleValue)
+				.toArray());
 		return this;
 	}
